Refactor Escrow table and conditionally render rows
